move system test to system tests directory
